requirements.txt
# Core Testing Framework
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-html>=3.2.0
pytest-timeout>=2.1.0

# AI/ML Libraries
openai>=1.0.0
anthropic>=0.7.0
transformers>=4.35.0
torch>=2.1.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Model Evaluation
rouge-score>=0.1.2
nltk>=3.8.1
sacrebleu>=2.3.1
evaluate>=0.4.0

# Bias & Fairness
fairlearn>=0.9.0
aif360>=0.5.0

# Performance Testing
locust>=2.15.0
memory-profiler>=0.61.0
psutil>=5.9.0

# Data Management
datasets>=2.14.0
pyyaml>=6.0
jsonschema>=4.19.0

# Reporting & Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
jinja2>=3.1.2

# API & HTTP
requests>=2.31.0
httpx>=0.25.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
tenacity>=8.2.0
